+++ examples/ppo_stable_baselines_idf_2023/__init__.py
@@ -1,30 +1,12 @@
 import os
 import json
 from grid2op.Agent import BaseAgent
-from .C_evaluate_trained_model import load_agent, agent_name
-from .gymenv_custom import GymEnvWithRecoWithDN as gymenv_class  # we use it here because l2rpn_baselines is not updated on the competition docker
-# from l2rpn_baselines import import GymEnvWithRecoWithDN as gymenv_class  # use this for the main tracks
+from .C_evaluate_trained_model import load_agent, agent_name, gymenv_class
 
 
 # two meta parameters not tune but that can be modified at inference time
 safe_max_rho = 0.99
 curtail_margin = 30.
-
-
-# the class of agent used during the tests
-class BaselineAgent(BaseAgent):
-  def __init__(self, l2rpn_agent, curtail_margin=150.):
-    self.l2rpn_agent = l2rpn_agent
-    self.curtail_margin = curtail_margin
-    BaseAgent.__init__(self, l2rpn_agent.action_space)
-  
-  def act(self, obs, reward, done=False):
-    action = self.l2rpn_agent.act(obs, reward, done)
-    # We try to limit to end up with a "game over" because actions on curtailment or storage units.
-    # this is "required" because we use curtailment and action on storage units
-    # but the main goal is to 
-    action.limit_curtail_storage(obs, margin=self.curtail_margin)
-    return action
 
 
 # the required function for codanbench